% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collect.hagstofa}
\alias{collect.hagstofa}
\title{Download a specified hagstofa object}
\usage{
\method{collect}{hagstofa}(d)
}
\arguments{
\item{d}{A hagstofa object defined using hg_data(url)}
}
\value{
Returns a tibble containing the data
}
\description{
Download a specified hagstofa object
}
