% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collect}
\alias{collect}
\title{Define the method collect that we use to download hagstofa objects}
\usage{
collect(x, ...)
}
\arguments{
\item{x}{}
}
\description{
Define the method collect that we use to download hagstofa objects
}
