% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_data}
\alias{download_data}
\title{Download a hagstofa object that contains < 5000 lines of results}
\usage{
download_data(d)
}
\arguments{
\item{d}{A hagstofa object defined using hg_data(url)}
}
\value{
Returns a tibble containing the data
}
\description{
Download a hagstofa object that contains < 5000 lines of results
}
