% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_large_data}
\alias{download_large_data}
\title{Download a hagstofa object that would give > 5000 lines.}
\usage{
download_large_data(d)
}
\arguments{
\item{A}{hagstofa object defined using hg_data(url)}
}
\value{
Returns a tibble containing the data
}
\description{
The data is downloaded by finding which variable uses the least amount of sub-queries and makes all sub-queries return < 5000 lines.
}
