% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tables.R
\name{hg_list_tables}
\alias{hg_list_tables}
\title{Get an interactive view of all available table from Hagstofa Íslands}
\usage{
hg_list_tables()
}
\value{
A searchable table made with the {gt} package
}
\description{
Get an interactive view of all available table from Hagstofa Íslands
}
